move query context into a higher-order component
